move nie-os app modules into folder for apps
