+++ src/app/fact-sheet/fact-sheet/fact-sheet.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { FactSheetComponent } from './fact-sheet.component';
 import { RouterTestingModule } from '@angular/router/testing';
-import { BaseTypeFormsModule } from '../../base-type-forms/base-type-forms.module';
+import { BaseTypeFormsModule } from '../../nie-OS-apps/shared/base-type-forms/base-type-forms.module';
 
 describe('FactSheetComponent', () => {
   let component: FactSheetComponent;
