+++ src/app/nie-OS-apps/create-resource/create-resource.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { BaseTypeFormsModule } from '../base-type-forms/base-type-forms.module';
+import { BaseTypeFormsModule } from '../shared/base-type-forms/base-type-forms.module';
 import { CreateResourceComponent } from './create-resource/create-resource.component';
-import { KnoraV1RequestService } from '../shared/knora/knora-v1-request.service';
+import { KnoraV1RequestService } from '../../shared/knora/knora-v1-request.service';
 
 
 @NgModule({
