+++ src/app/nie-OS-apps/image-frame/image-frame-sizes/image-frame-sizes.component.ts
