+++ src/app/nie-OS-apps/image-frame/region-to-svg.service.ts
