+++ src/app/nie-OS-apps/resource-form/resource-form/resource-form.component.spec.ts
@@ -3,12 +3,12 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { HttpClientModule } from '@angular/common/http';
 
-import { BaseTypeFormsModule } from '../../base-type-forms/base-type-forms.module';
+import { BaseTypeFormsModule } from '../../shared/base-type-forms/base-type-forms.module';
 
 import { ResourceFormComponent } from './resource-form.component';
 import { ResourceValueHistoryComponent } from '../resource-value-history/resource-value-history.component';
 import { ResourceValueHistoryValueComponent } from '../resource-value-history-value/resource-value-history-value.component';
-import { KnoraV1RequestService } from '../../shared/knora/knora-v1-request.service';
+import { KnoraV1RequestService } from '../../../shared/knora/knora-v1-request.service';
 
 describe('ResourceFormComponent', () => {
   let component: ResourceFormComponent;
