+++ src/app/nie-OS-apps/shared/base-type-forms/int-value-editor/int-value-editor.component.ts
