+++ src/app/nie-OS/nie-OS.module.ts
@@ -15,7 +15,7 @@ import {MatTooltipModule} from '@angular/material/tooltip';
 import {SynopsisModule} from '../synopsis/synopsis.module';
 import {MyPageSetModule} from './page-set/page-set.module';
 import {GenerateHashService} from '../shared/nieOS/other/generateHash.service';
-import {CreateResourceModule} from '../create-resource/create-resource.module';
+import {CreateResourceModule} from '../nie-OS-apps/create-resource/create-resource.module';
 import {DataManagementToolModule} from './apps/data-management-tool/data-management-tool.module';
 import {SendGravSearchQueryService} from '../shared/knora/gravsearch/sendGravSearchQuery.service';
 import {TextlistViewerComponent} from './apps/textlist-viewer/textlist-viewer.component';
